Fix yara-x build in QEMU (#38328)
I noticed that running `make package/yara-x` would fail when using the
QEMU runner.

This is due to `cargo install cargo-c` leveraging `/tmp` which is
mounted as `noexec` in the QEMU runner:
https://github.com/wolfi-dev/os/blob/2825e0b3c582679cef961cd97b6b7cc47cb50516/melange/init#L10

To comply with this restriction, this PR sets the installation root for
`cargo-c` to `$HOME/.cargo` rather than `/tmp`.

Signed-off-by: egibs <20933572+egibs@users.noreply.github.com>