Build yara-x with increased memory, --target-dir for cargo-c, and optimizations (#38392)
I had some local Melange changes that made #38328 appear valid when in
fact the build artifacts for `cargo-c` were still landing in `/tmp`.

This PR specifies `--target-dir` to redirect `cargo-c`'s build artifacts
to a non-`/tmp` directory to avoid running in to QEMU's `noexec` tmpfs.

Once this was fixed, the `cargo-c` install was OOM'ing with the default
`-m 4000000k` configuration so I added a memory requirement of `6Gi`
which allowed the build to succeed. Finally, I added `opt-level=3` to
build the `yara-x-capi` with all available optimizations.

---------

Signed-off-by: egibs <20933572+egibs@users.noreply.github.com>