nvidia-container-toolkit: fold -operator into -nvidia-toolkit
The -operator package provides /usr/bin/operator, which is
not actually an executable. It is a compiled go package object
that is built into the nvidia-toolkit binary. This has always
been the case, as can be observed by installing the initial
version released in wolfi[*]. Let's stop building it, and
add a provides for it to -nvidia-toolkit to prevent
accidentally shipping an old stale version.

[*]
  # apk add nvidia-container-toolkit-operator=1.15.0-r0
  fetch https://packages.wolfi.dev/os/x86_64/APKINDEX.tar.gz
  (1/1) Installing nvidia-container-toolkit-operator (1.15.0-r0)
  Executing busybox-1.37.0-r0.trigger
  OK: 14 MiB in 16 packages
  # /work/operator
  /bin/sh: /work/operator: Permission denied
  # file /work/operator
  /work/operator: symbolic link to /usr/bin/operator
  # file /usr/bin/operator
  /usr/bin/operator: current ar archive

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>