use cargo/build pipeline and fix build failure for eza package
this was a little unusual because I was not able to find anything
concrete in logs.

I tried building it locally and was facing conflicts with git2 b/w rust
and libgit2 packages.

I removed libgit2-dev from build time dependencies and then I used
cargo/build pipeline which uses cargo-auditable to build eza. This is
because we can use rust-audit-info to get the crates that was used to
build this binary.

Signed-off-by: kranurag7 <81210977+kranurag7@users.noreply.github.com>