ruby3.* activemode/support
activemodel requires a matching version of active support and should be updated
at the same time ideally in the same PR. Upstream will release them at the same time.
I'm adding a note to remind people.  Our packages will install fine but not be
able to import.  This will show up as a runtime failure for testing activemodel
but not when testing activesupport.

https://github.com/rails/rails/blob/be9aa73dd72f1097be5d45a58d7912447a266bd1/rails.gemspec#L32