remove package `up`, update dependents to fetch prebuilt `up` (#38724)
The up CLI seems to have been made closed-source:

- github.com/upbound/up 404s
- [Official docs](https://docs.upbound.io/reference/cli/) make no
mention of installing from source

Unfortunately, `up` is still required to build some official Crossplane
providers (AWS, GCP, etc.), so rather than using our
increasingly-out-of-date and vuln-ridden `up`, we'll hold our nose and
install a pre-built closed-source `up`.

This is only a build-time dependency, at least, so it won't end up in
any of the packages or images produced as a result.

supercedes
- https://github.com/wolfi-dev/os/pull/38140

---------

Signed-off-by: Jason Hall <jason@chainguard.dev>