fix helix build failure
helix package was failing because of cargobump. Some dependencies were
conflicitng/missing.

I removed cargobump because the CVEs is fixed in upstream project
already.

this commit also removed build-time dependencies because we no longer
need that. We're using cargo/build pipeline which pulls in the right
packages required for building helix.

Signed-off-by: kranurag7 <81210977+kranurag7@users.noreply.github.com>