gcc-11: sync packaging with gcc-12
Merge most of the delta between gcc-11.yaml and gcc-12.yaml.
I manually backported the patch because it looks like it
would still be needed.

The remaining delta - other than s/11/12/ is as follows,
with comments inline.

> --- /tmp/gcc-11.yaml
> +++ /tmp/gcc-12.yaml
> @@ -1,7 +1,7 @@
>  package:
>    name: gcc-11
> -  version: 11.5.0
> -  epoch: 1
> +  version: 12.4.0
> +  epoch: 7

> @@ -24,13 +24,13 @@
>        - ca-certificates-bundle
>        - flex-dev
>        - gawk
> -      - gcc-12-default

We're building gcc-11 w/ gcc-12, that's OK.

>        - mpfr-dev
>        - openssf-compiler-options
> +      - patch

The patch pipeline will add this.

>    - name: 'libstdc++-11-dev'
>      dependencies:
>        runtime:
> -        # cargo-culted from gcc-12.yaml
> +        # libstdc++-11-dev is no longer compatible at link time with
> +        # libstdc++ 13, despite being runtime compatible Ensure any
> +        # users of libstdc++-11-dev have access to link against
> +        # libstsdc++-11
>          - libstdc++-11

Unsure if this is necessary, but it shouldn't hurt. I just changed the
comment.

>      pipeline:
>        - runs: |
> @@ -168,7 +176,7 @@
>      description: 'Use GCC 11 as system gcc'
>      dependencies:
>        provides:
> -        - gcc=${{package.version}}
> +        - gcc=12.3.1

I'm not sure why we hardcoded 12.3.1 for gcc-12. This
seems sufficient for gcc-11.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

foo