systemd: Add container package for container related tools and daemons (#38673)
This new package contains the follow pieces used for container and image
management which are not needed for most systems.

Notably sysext and context are being left in the main systemd package
but portablectl is being included here. Debian puts it in their
system-container package but fedora doesn't. I think it fits well here
so that's where I put it.

machine
portable
systemd-disect
vm/nspawn
import/export

---------

Signed-off-by: Luca Di Maio <luca.dimaio1@gmail.com>
Co-authored-by: Justin Vreeland <justin.vreeland@chainguard.com>
Co-authored-by: Luca Di Maio <luca.dimaio1@gmail.com>