fix(cert-manager): Force rebuild of packages and images following image fix
PR https://github.com/wolfi-dev/os/pull/34777 introduced a package rename from cert-manager-1.16-acmesolver -> cert-manager-acmesolver-1.16

This rename was not reflected in the image defintion causing the image to continue to pull in the older packages -
these older packages are now vulnerable to CVEs.

This commit bumps the epoch to force a new package version and thus a new image build.

Signed-off-by: philroche <phil.roche@chainguard.dev>