mysql-9.1: create default runtime directories during build
`mysqld` won't start in the default configuration which we ship without
these directories being present; this commit avoids users having to
discover the issue and create them themselves.

To validate this change, this commit also updates the test pipeline to
run with the default value for --datadir (which is the same as what we
were passing anyway), and not to create the directories which the
package now ships.