py3-cassandra-medusa: drop idna version override
We bumped idna to 3.7 in commit aebb92a29dfe021, for upstream
version 0.20.1 to "fix CVE". I'm guessing this was to fix:
  https://github.com/advisories/GHSA-jjg7-2v4v-x38h

Upstream pins idna at >= 3.6, but I've verified that we get
3.7 anyway. Avoid the upstream override so that we can
inherit dependency changes from upstream without manual
intervention.

I've asked upstream to bump their pin to 3.7 for the sake
of correctness:
  https://github.com/thelastpickle/cassandra-medusa/pull/848

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>