py3-cassandra-medusa: fix requests version override
We added an override to bump `requests` to version ^2.23.0
in commit 9e6288702c8bc87e on upstream version 0.21.0.
This was to fix CVE-2024-35195, which `requests` upstream fixed
in version 2.32.0.

I'm guessing this was a simple character transposition
(2.23 instead of 2.32). But by unlocking us from upstream's
pinned 2.31.0, it did the trick. Today it installs 2.32.2.

But let's correct the version, and leave a note for future
updaters.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>