Fix for update-bot failing to poll repo - also flips from fork to main source repo (#39472)
Fix for update-bot failing to poll repo - by migrating to git poller.

This package was also referencing a fork, and that fork does not contain
any additional changes vs original upstream repo, not does it contain
any release tags. This seems... incorrect? Perhaps not intended
originally. I bumped the epoch to trigger a build given we flipped the
source repo.

---------

Signed-off-by: Mark McCormick <mark.mccormick@chainguard.dev>