Update Git checkout URL (#39592)
This package currently leverages release-monitor and fetch to pull down
the source code. The URL has the year hard-coded, which needs to be
updated.

Longer term we might consider moving this to a git clone, but the
reasons for not doing so originally would need to be understood. It's
sometimes intentional for larger projects

---------

Signed-off-by: Mark McCormick <mark.mccormick@chainguard.dev>