chromium: don't use system libdrm
Support for building with system libdrm was dropped on Linux in
https://github.com/chromium/chromium/commit/a20ac354f56c8a55a4e5b4690753f3c08b7d0067

This upstream change isn't strictly necessary with a modern libdrm: the
commit message mentions working around the absence of the
`drmSyncobjEventfd` symbol, but that was introduced to libdrm in
https://gitlab.freedesktop.org/mesa/drm/-/commit/c6013245ce9ce287bb86d327f9b6420a320a08e6
and has been in upstream releases for over a year.  Reverting the
above-mentioned commit results in a successful build using the system
libdrm.

However, to avoid getting out of step with Chromium upstream (who vendor
libdrm and could therefore choose to carry patches to it which _aren't_
available from an upstream libdrm release), and because the source code
itself is in the Chromium repo (and therefore patchable if a CVE
requires it), we'll stop using our own libdrm for these builds.