chrony: disable NTS, MAC, MS-NTP unused or requires bootstrap (#39793)
NTS, pre-shared MAC keys, MS-NTP allow to configure authenticated time
sources, but they require bootstrap of having accurate time to begin
with (to verify TLS certificates NotBefore/NotAfter) or it requires
sharing symmetric keys (which is impractical for public network
sources).

This is widely unused, as there is no specification on supporting this
with pools, and in public clouds local PTP, local time source over
trusted network, and finally falling back to public time sources is
the current mode of operation.

If need for these protocols arises, we can build a separate
sub-package with support for NTS, MAC, MS-NTP.

This removes 9 runtime dependencies.