Add yara-x patch to reduce virtual memory usage (#39901)
This PR adds a yara-x patch to handle WASM memory management tweaks. By
default, `wasmtime` is allocating 8GB of contiguous virtual memory each
time a new scanner is created. This is fine except in cases where the
file-system is in-memory or other virtual memory restrictions are
present (e.g., in Cloud Run).

To avoid this, we can use dynamic allocation and drop the overall
allocation and grow usage by a lower amount.

I specified `config.static_memory_guard_size` because its value is `2GB`
by default and it cannot be lower than the default dynamic guard size of
`64KB`.

I also added another test case to make sure that malcontent runs with
the updated code. The real test will be running the updated API in an
environment similar to Cloud Run.

---------

Signed-off-by: egibs <20933572+egibs@users.noreply.github.com>