New package rustfilt
This package is similar to c++filt for c++, but for rust. It helps to
demangle symbols names from gibberish to human readable function
names. This helps with static analysis of rust binaries, for example
to determine which crates got linked into a binary.

Packaging git snapshot to get this feature:
- https://github.com/luser/rustfilt/commit/4618d5f5a00e7c7cb6d71c5851c80e18f53a5509

Until the next upstream release is cut:
- https://github.com/luser/rustfilt/issues/23