py3-conda-libmamba-solver: update test-comprehensive for multiprocessing
This test was generating error output because execution of it uses
multiprocessing, which (a) requires an on-disk file to import in each
process, and (b) will execute anything at the module-level in each
process.

This commit writes the script to disk, and moves all its behaviour into
`main()` which is called only by the `__main__` process (and not other
importing processes).