zetasql: strip test files with non-latin filenames
boost and civetweb include test files with non-latin filenames.
Something in the build system is unable to deal with them:

 WARN ERROR: /home/build/.cache/bazel/_bazel_build/79a1bfc8c8b5b6a6d226b38d072f165b/external/boost/BUILD.bazel:27:12: Foreign Cc - BoostBuild: Building boost failed: error reading file '@boost//:libs/wave/test/testwave/testfiles/utf8-test-ßµ™∃/file.hpp': /home/build/.cache/bazel/_bazel_build/79a1bfc8c8b5b6a6d226b38d072f165b/external/boost/libs/wave/test/testwave/testfiles/utf8-test-ßµ™∃/file.hpp (No such file or directory)
 WARN ERROR: /home/build/.cache/bazel/_bazel_build/79a1bfc8c8b5b6a6d226b38d072f165b/external/boost/BUILD.bazel:27:12: Foreign Cc - BoostBuild: Building boost failed: 1 input file(s) are in error

Note that it is not the project test suites detecting a regression,
rather it is part of the bazel build setup.

Until this is root caused, work around it by stripping those
files out of the build area.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>