Fix(Build):Fix pipeline for broken build
Feat(test): add test for the binary

The previous process used  to package the source code into a tar file, which is not required for our package.

Starting from version 7, the  target has been removed from the Makefile. Now, the application is built and stored inside the  directory, where the  binary can be found.

Signed-off-by: Debasish Biswas <debasishbsws.dev@gmail.com>