build ruby linked against jemalloc (#40144)
this changes all our ruby to be linked against `jemalloc`. for
applications like webapps (rails) or multithreaded workloads (puma), we
see significant memory usage savings compared to malloc.

I don't know enough as to whether we should do this globally, since this
introduces some size overhead to our binaries, so I'm going to leave it
to the experts to yay or nay this.

if nay, we'd likely need a jemalloc enabled subpackage or something
similar just for the sheer perf gains we see against any rails/puma app

---------

Signed-off-by: Josh Wolf <josh@wolfs.io>
Signed-off-by: dann frazier <dann.frazier@chainguard.dev>
Co-authored-by: dann frazier <dann.frazier@chainguard.dev>