pipelines: Add compiler/filter-openssf-flags
There are some compiler flags set by openssf.spec that
can't be trivially removed. `-Wl,-z,now` is one such
example, for which a copy and edit hack has been
propagating. Let's create a pipeline for this so
that we only have to maintain that code in one place.

Use of this should be discouraged if other options
are available. But if the decision is between using
this or disabling all openssf hardening flags, this
is the better choice.

Signed-off-by: dann frazier <dann.frazier@canonical.com>