move xvfb-run to a separate package
separate out xvfb-run subpackage into it's own package because it was
using a separate source and upon new commit the sha validation was
failing leading to build time errors.

one solution was to stop validation commit sha and clone using branches.

this moves to a package package and we now poll the branches weekly
using git backend. bot will keep on bumping it weekly if upstream
receives any new commit.

Signed-off-by: kranurag7 <81210977+kranurag7@users.noreply.github.com>