py3-google-cloud-pubsub: withdraw package and broken versions
This package is no longer used for image builds and the lack of tests
mean that it has bitrotted since its last active usage.  The package
isn't configured for multiple versions of Python, and now depends on
other Python modules which we don't have packaged, so fixing it is more
work than an unused package justifies.  If in future we have demand
for it, we can reintroduce it "properly" then.

This also withdraws versions from 1.25.0-r0 onwards: upstream introduced
a new dependency in 1.24 (a version we never released) imported at
module-level, making the module entirely unimportable in those versions
and later.  Previous versions are left in place (despite lacking
multi-Python support).