Makefile: Declare .PHONY targets
I somehow ended up with a `local-wolfi` directory in my build tree,
likely a fat-fingered command, and found `make local-wolfi` didn't
work because make though it was already up to date :)

There's a syntax for telling make that the target doesn't actually
generate a file (or directory) of that name. Use it for all such
targets.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>