glibc: Re-enable OpenSSF hardening
glibc was failing to build with openssf-compiler-options
on ARM because those options enable branch protection (BTI/PAC).
This feature can only work if all objects are built with
BTI/PAC, and that was not the case with the GCC code that
gets linked into glibc. GCC has been fixed, so let's
restore hardening.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>