apache2: Add runtime dependency on busybox
The new entry point is a script that runs `rm`. We had an image fail
to test because the `rm` command did not exist, so add a dep on `busybox`.

`busybox` also provides a shell, so we can drop the `dash-binsh` dependency.
The `dash-binsh` dependency was added along with the previous entrypoint
script in commit 8dad232d60bca144b357cb652cb0865d6e351be7, so I conclude
it is just there to handle the entrypoint, which busybox's shell handles
fine.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>