bump mc package to rebuild bitnami compat
removes 2024 from the package name to simplify things at image level.
This way we'll only require to unblock the package without any changes
at image level.

keeps update to true with a note to check the URL, this way it'll fail
and during fixing the package we'll update the URL.

Signed-off-by: kranurag7 <81210977+kranurag7@users.noreply.github.com>