opa-envoy: combine mangled-package-version var-transforms
melange will always write to the `to` of `var-transforms`, even if the
`from` value did not `match` and so has not been mutated.  This means
that the last `var-transform` with a given `to` will always determine
the value of that variable.  In this case, that meant that our
`var-transforms` would only ever handle one of the two upstream
versioning cases.

This commit collapses the two stanzas into a single (ugly) regex which
should handle both cases, and I've filed
https://github.com/chainguard-dev/melange/issues/1774 to request making
the multi-stanza version possible.