glibc: add test for compatibility with gcc (#42242)
Add the test for the issue we missed when trying to release 2.41
prematurely. There is a very very very low chance this will happen
again, but I still think it's a good idea to add a test. In case
something similar is broken with pthreads once again.

This can land as-is, doesn't actually require the new glibc. Confirmed
to be failing with a straight rebuild of glibc without fixing gcc
fix-includes.

Signed-off-by: Łukasz 'sil2100' Zemczak <lukasz.zemczak@chainguard.dev>