java-cacerts: update to 20241121 bundle (#42291)
Update java-cacerts to 20241121 bundle, also attempt to enable updates
to it, with lockstep against ca-certs.

If this doesn't work, we should just rebuild/refetch ca-certificates
in java-cacerts from scratch. Due to build-dependency cycles we can't
do it all in ca-certificates itself, but we should be able to redo it
again.