abseil-cpp: merge all libs into a single package
The one-package-per-lib split may cause more issues than it solves.
Each one provides an SBOM, which is sometimes larger than the lib
itself, so it isn't always a space saver. Scanners are going to
report CVEs at the source package level, so we're not reducing
that hit rate. And this adds more opportunities for something
to fail in our infrastructure.

See the discussion here:
  https://github.com/wolfi-dev/os/pull/41278#issuecomment-2668523420

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>