fix(calico-3.29.yaml): Disable the calico-apiserver 3.29.2 when Validating Admission Policy checks until upstream tigrea operator adds the required rules (#43188)
Because we are using tigera-operator during image test we are reliant on
tigera-operator setting
up all the required rbac rules that are needed for calico-apiserver to
run. Currently it does not [1] and
this is a temporary solution until
https://github.com/tigera/operator/issues/3780 is resolved upstream with
a new
release of tigera-operator. This issue was highlighed in
calico-apiserver 3.29.2 when Validating Admission Policy
became enabled by default [2]. This patch disables the Validating
Admission Policy in calico-apiserver until
the issue is resolved.

[1]
https://github.com/tigera/operator/blob/ed26c42e1e232e9cc604a9a010d4ca9111b6f388/pkg/render/apiserver.go#L654C11-L654C28
[2]
https://github.com/projectcalico/calico/commit/8eacb02e223edeef01e28d85df794540c6850d0f

Signed-off-by: philroche <phil.roche@chainguard.dev>

Signed-off-by: philroche <phil.roche@chainguard.dev>