rsyslog - set runstatedir to /run , and PIDFILE to /run
The symlink here was causing issues on installation via apko,
and it is clearly not right for rsyslog to own that.

--runstatedir is recognized by rsyslog and defaults to /var/run.
However, that actually does not affect where the PIDFILE gets
written.

So we have to set PID_FILE_PATH in the environment to accomplish
the change of pid file location.

There is a test cleanup also. Instead of running a 'rsyslogd -N1'
to check the config after daemon-start-stop, run that as its
own thing.