daemon-check-output - do not just exit failure if kill fails. (#43328)
We would not expect for 'kill' to fail here, but becauase of 'set -e'
execution if it does we don't catch that error. The logs clearly
expected to continue on.

I did actually see this failure path in test of rsyslog in bwrap. The
daemon would start and be running, but 'kill' would fail with permission
denied.

The change here just makes the script continue as it was obviously
expecting to do.