Adds compat package to provide compatibility with Bitnami helm chart (#43785)
Adds a new subpackage to provide compatability with the Bitnami helm
chart for metrics-server. Note at present time, there aren't any custom
scripts or entrypoints required for this compatibility, other than the
binary being symlinked in the directory that the helm chart expects, and
some configuration at the image level.

Thus this is a pretty straight forward subpackage.

Signed-off-by: Mark McCormick <mark.mccormick@chainguard.dev>