keycloak-config-cli: fix bitnami symlink (#44312)
The Bitnami keycloak-config-cli image defaults to the `bitnami - 1000:0`
user, however the defaults in their helm chart runs the container as
`bitnami - 1001:1001` which throws an error about not being able to
access the jarfile referenced by the current symlink.

This fix changes the symlink to the full filepath and adjusts the
subpackage test to confirm that this runs with the default Bitnami helm
chart uid:gid.

Signed-off-by: Michael Paul <31833332+MickeyPvX@users.noreply.github.com>