git: enable ELF package notes canary (#44221)
git is present in all -dev tags, thus as a single package this can
affect roughly half of our image tags. Many users do use and scan -dev
tags. A single package in -dev tags with ELF notes is a very good
canary, that will give us a good signal if this is being tolerated by
our user base and scanners. And also, it is very easy to revert, as it
is a single package withdrawal & rebuild.