gobject-introspection - separate binaries to -bin package. (#43542)
The change here allows someone to use the library
libgirepository-1.0.so.1.0.0
(so:libgirepository-1.0.so.1=1) without incurring the cost of of the
runtime
dependencies of the executables that it ships. This is accomplished by
moving
those inaries to a -bin package.  executables are:

g-ir-annotation-tool g-ir-compiler g-ir-generate g-ir-inspect
g-ir-scanner

The -dev package now depends on the -bin package, so consumers of the
-dev
package will not be affected.

Things that previously listed just the runtime 'gobject-introspection'
but used
any of the executables will need to add gobject-introspection-bin but
should
probably add -dev.