mariadb-11.7 - Include a configuration file with appropriate defaults
Post /bin usrmerge, /bin/mariadbd now exists because of the /bin symlink to /usr/bin, where mariadbd actually resides.
The mariadb docker entrypoint startup scripts check for the existence of /bin/mariadbd, and guess that the installation root is / when it is actually /usr, causing a failure to find some startup SQL scripts.