Update Spark package configuration for OpenJDK and enhance PySpark di… (#44730)
Coming from this
[escalation](https://github.com/orgs/chainguard-dev/projects/45/views/62?pane=issue&itemId=100254391&issue=chainguard-dev%7Ccustomer-issues%7C2097)
customer was unable to use kubernetes libraries in the
`spark-3.5-scala-2.13` package.
This change adds the `-PKubernetes` build parameter for building a spark
distribution.

Also used this change to rename the `spark-3.5-minimal-openjdk-` package
to `spark-3.5-scala-2.12-minimal-openjdk-` to match what is expected in
the custom package
[here](https://github.com/chainguard-images/images-private/blob/main/images/request-471/main.tf#L5)

Update Spark package configuration for OpenJDK and enhance PySpark
distribution script

---------

Signed-off-by: uti <uti.edun@chainguard.dev>