fix(selenium): Use correct versions of things (#44834)
Docker Selenium should always be built with the versions of gRPC, Netty,
and OTel consumed by upstream (unless we've a compelling reason not to,
i.e. vulnerability remediation). This addresses issues encountered when
tracing is enabled. Likewise, the version of DS should always be
consistent with the version of Selenium. The two should never be
mismatched

Also land Selenium 4.29.0 and move on to Java 21 per upstream

---------

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>
Co-authored-by: Daniel Watkins <daniel.watkins@chainguard.dev>