apk-tools: re-enable update bot
v2.14.3 had an incompatibility with Wolfi OS and it was withdrawn from
Wolfi, and updates disabled, see:
- https://github.com/wolfi-dev/os/pull/16207

But since then, we successfully upgraded to v2.14.4 in:
- https://github.com/wolfi-dev/os/pull/21906

But since then, we didn't turn the updates back on. Let's turn them
back on.

This PR will cause bot to open a PR with these upstream changes:
- https://github.com/alpinelinux/apk-tools/compare/v2.14.4...v2.14.10