add a real test for pod_coverage (#46122)
perl-pod-coverage installs a perl module and wrapper script that tests
documentation coverage in Perl modules. Add a test that installs a
package from Wolfi (openssl-dbg) that happens to ship a Perl MANIFEST,
and run pod_coverage on it. Note that if, at some point in the future,
openssl-dbg doesn't ship that MANIFEST, then this test will need to
fetch a module from elsewhere. Sure, whatevs.

Co-authored-by: Dustin Kirkland <dustin.kirkland@chainguard.dev>