add tests for bazelisk, fix runtime dependency problem (#45792)
Add a couple of tests for bazelisk, which actually identified a missing
runtime dependency on libstdc++.

Why didn't SCA catch this? Perhaps because the first thing this binary
does is reach out to the internet and fetch a newer copy of itself? Is
that what we want? Not sure... If we want this to work out of the box in
this manner, then we need to add the runtime dependency on libstdc++. On
the other hand, a binary that reaches out to the internet and downloads
a different thing to run doesn't really feel like the secure source for
open source....

Definitely need a review on this one....

Co-authored-by: Dustin Kirkland <dustin.kirkland@chainguard.dev>