add busybox to spark (#45910)
`spark-3.5-scala-2.13` was recently compiled with `-Pkubernetes` to
allow deploying spark in a kubernetes cluster. The jars needed for this
are available in the package and therefore in the image, the absence of
a shell makes it difficult to verify this in a non dev docker image.
This change adds busybox as a runtime dependency to the package.