clean(wolfi-baselayout): cleanup xfs deprecated user and group
The X Font Service user account is deprecated. Alpine baselayout already
cleaned up their user and groups a year ago [1]. This is following its
approach in a gradual way, by beginning with users which are deprecated
and that generates uid conflicts in passwd database. In example, the uid
33 is usually associated to web services, like www-data in Debian or
http in Arch.

1.
https://gitlab.alpinelinux.org/alpine/aports/-/commit/f16d0754d60196f00c3fad7f881d3a9bd6943152

Signed-off-by: Massimiliano Giovagnoli <massimiliano.giovagnoli@chainguard.dev>