nodejs*/python*: Workaround issue with rosetta2 and sahf
There seems to be an issue when running code with the sahf instruction
on rosetta2.  We first encounterd this with nodejs and used
`-march=x86_64` to fix it. We have since encountered more instances and
a reproducer with python.  `python3 -c 'print((-0.01)**2);'` When compiled
will cause the same python to hang basically forever maxing out a cpu.