add docker runner specific make targets (#48092)
some packages both during build and test times behave different when we
use docker runner and when we use bubblewrap runner.

We use bubblewrap on ws and mac and docker in CI. Adding these make
targets so that it's easier to have access to docker specific runners.

less typing and I continue to be lazy :)

Signed-off-by: kranurag7 <81210977+kranurag7@users.noreply.github.com>
Co-authored-by: Wojciech Kocjan <wojciech.kocjan@chainguard.dev>