Add cluster-api (#48164)
Add cluster-api meta package version 1.9.6 with subpackages and
pipelines
The package requires a full blown kubernetes cluster for complete
testing so most of the tests are just to test the execution of the
binaries for each component

<!---
Provide a short summary in the Title above. Examples of good PR titles:
* "ruby-3.1: new package"
* "haproxy: fix CVE-2014-123456"
-->
cluster-api-1.9: new package
<!--
Please include references to any related issues or delete this section
otherwise.
 -->

Fixes:

Related:
#### For new package PRs only
<!-- remove if unrelated -->
- [x] REQUIRED - The package is available under an OSI-approved or
FSF-approved license
- [x] REQUIRED - The version of the package is still receiving security
updates

---------

Signed-off-by: uti <uti.edun@chainguard.dev>