unbound Add some extras to mailcow-compat
Provide configuration for supervisord and syslog-ng.

Move entrypoint and healthcheck for containers to be aligned
to mailcow/unbound OCI image.

Rename -conf -> -config and adjust priority around default of 0 so
that alternative configuration files can be switched in without
broader impact.

Add additional runtime dependencies for compat package to
support heathchecks and logging.

Signed-off-by: James Page <james.page@chainguard.dev>