openssf-compiler-options: Provide linker flags via link spec
We're currently providing linker flags using the -Wl,... mechanism via
the "self_spec" definition.  This works, but confuses GCC when no
input files have been provided (as is the case when running "gcc -v",
for example).  GCC thinks that there's actually something being
compiled, and then errors out when it finds that there's no "main"
function defined.

The spec file has a better way to specify linker flags: we can append
to the "link" spec and pass the flags directly.  It fixes the problem
with "gcc -v" while still DTRT when linking programs.

Closes: #48510

Signed-off-by: Sergio Durigan Junior <sergiodj@chainguard.dev>