glib: bump to correct elastic-build ordering
`libgsf` depends on `glib` which in turn depends on `libxslt` and
`gettext`.  `glib` itself does not require a bump for the new `libxml2`
soname, but if it isn't in the set of packages being built,
elastic-build won't correctly order `libgsf`'s build after the `libxslt`
and `gettext` builds, resulting in a failure to create its build
environment (as it uses the unbumped `libxslt`/`gettext` packages which
depend on the soname in libxml2 2.13.6-r1, meaning 2.14.0-r0 can't be
added into the dependency graph).