add subpackages and use package.version in testing
the tests were failing because we were hardcoding the version inside the
tests.

moved that to use ${{package.version}} meaning it'll pass on every
update automatically.

the main package was also including docs and locales so moved that to
separate subpackages making the size of main package a little bit more
smaller.

Signed-off-by: kranurag7 <81210977+kranurag7@users.noreply.github.com>