Patch GHSA-3wgm-2gw2-vh5m in our k8s package by disabling git repo pl… (#49254)
…ugin.

This has been deprecated for 6 years and the vulnerability isn't getting
fixed. It's due to be disabled by default in 1.33, this forwards that
up.

Fixes:

Related:

### Pre-review Checklist

<!--
This checklist is mostly useful as a reminder of small things that can
easily be
forgotten – it is meant as a helpful tool rather than hoops to jump
through.

At the moment of this PR you have the most information on what all the
change
will affect, so please take the time to jot it down.

Put an `x` in all the items that apply, make notes next to any that
haven't been
addressed, and remove any items that are not relevant to this PR.

-->

#### For security-related PRs
<!-- remove if unrelated -->
- [ ] The security fix is recorded in the
[advisories](https://github.com/wolfi-dev/advisories) repo

#### For version bump PRs
<!-- remove if unrelated -->
- [ ] The `epoch` field is reset to 0

#### For PRs that add patches
<!-- remove if unrelated -->
- [ ] Patch source is documented

Signed-off-by: Dan Lorenc <dlorenc@chainguard.dev>