linkerd-extension-init: modify tests
Linkerd2 now requires the kubernetes gateway-api CRD to be enabled by
default.
We now apply the CRD from kubernetes before installing linkerd.
We have also modified the Integration test as it seems that kwok loses
the context in between runs and needs to be restarted again, so we
setup, start and verify in a single run now.

Signed-off-by: David Negreira <david.negreira@chainguard.dev>